unify clients test
